+++ src/components/movie-details/movie-details.jsx
@@ -3,12 +3,14 @@ import PropTypes from 'prop-types';
 import MovieCardFull from "../movie-card-full/movie-card-full.jsx";
 import Footer from "../footer/footer.jsx";
 import CatalogLikeThis from "../catalog-like-this/catalog-like-this.jsx";
+import connect from "react-redux/es/connect/connect";
 
 const MovieDetails = (props) => {
-  const {film, films} = props;
+  const {films, filmId} = props;
+  const currentFilm = films.findIndex((film) => film.id === filmId);
 
   return <Fragment>
-    <MovieCardFull film={film}/>
+    <MovieCardFull film={films[currentFilm]}/>
 
     <div className="page-content">
       <CatalogLikeThis
@@ -19,9 +21,17 @@ const MovieDetails = (props) => {
   </Fragment>;
 };
 
-MovieDetails.propTypes = {
-  film: PropTypes.object,
-  films: PropTypes.arrayOf(PropTypes.object).isRequired
+MovieDetails.defaultProps = {
+  films: []
 };
 
-export default MovieDetails;
+const mapStateToProps = (state) => {
+  return {films: state.films};
+};
+
+MovieDetails.propTypes = {
+  films: PropTypes.arrayOf(PropTypes.object).isRequired,
+  filmId: PropTypes.number.isRequired
+};
+
+export default connect(mapStateToProps)(MovieDetails);
